Moved server.js into app.js, this abstraction was needed to run tests without starting the http server, also added first tests, and made some response changes for better error handling.
