+++ routes/emails.js
@@ -28,9 +28,10 @@ exports.sendVerificationEmail = async (recieverEmail, token) => {
 
     transporter.sendMail(mailOptions, function (error, info) {
         if (error) {
-            console.log(error);
-        } else {
-            console.log('Email sent: ' + info.response);
+            return error;
+        }
+        else {
+            return "success";
         }
     });
 };
